requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
PyYAML
Pillow
diffusers
matplotlib
requests
torchvision 
PyQt6